This patch implements a PyAuto-based integration test for WebRTC. The purpose is to do a simple end-to-end test for the new WebRTC component in Chrome.
Currently, the test is quite basic and only verifies that we can get through the API calls successfully. No video or audio is verified. This patch contains some changes to the Chrome automation controller so we can click past the media stream infobar (e.g. the 'is it ok for this webpage to use the camera and mic' dialog).

BUG=
TEST=

Review URL: http://codereview.chromium.org/10115019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133175 0039d316-1c4b-4281-b951-d872f2087c98