Assert when URLRequests leak for the main URLRequestContexts.
I've already enabled the assertions for all the non-main URLRequestContexts:
  * safe browsing
  * system
  * media
  * extensions

I've squashed those leaks, so now it's time to work on the main URLRequestContext for the profile. I know there are defintely existing leaks, since I see use-after-free crashes, so I expect to trigger a number of crashes here. I will revert this by Monday so as not to have this change in the dev channel release.

BUG=90971
TEST=none


Review URL: http://codereview.chromium.org/10139014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133313 0039d316-1c4b-4281-b951-d872f2087c98