Check RendererAccessibility mode instead of WebKit accessibilityEnabled.
We were hitting an assertion failure that indicated that WebKit accessibility
was enabled, but RendererAccessibility's mode was that accessibility is off.
Note that WebKit accessibility is a global flag, once enabled, it is never
turned off. This could happen if a new RenderViewImpl is created in the
same process where WebKit accessibility was previously enabled, but the
newly-created RendererAccessibility mode is off.

The fix is to never call handle accessibility notifications if
RendererAccessibility's mode is off, regardless of WebKit's accessibility
notification flag.

BUG=123649
TEST=none


Review URL: http://codereview.chromium.org/10162028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133489 0039d316-1c4b-4281-b951-d872f2087c98