Allow about:tracing Counters to be selected, with supporting refactorings.
The selection infrastructure in about:tracing was grown organically and in a hurry.
This refactors the organic selection logic around the concept of a TimelineSelection,
which represents an array of TimelineSelectionHit objects. A selection hit is
created by a track and contains enough information to uniquely identify the
selection both in the track and in the overall TimelineModel.

Counter tracks are then updated to support this new picking infrastructure,
generating TimelineSelectionHits. Whereas TimelineSelectionSliceHit stores its
selected state inside the timeline model, TimelineSelectionCounterHit stores
its selected state inside the track. The counter hits follow the new intended
model; a followon refactoring will be needed to clean up the TimelineSlice
selection state.

As part of this refactoring, the analysis code for traces is pulled from the
timeline view into a standalone analysis class. This allows cleanup and unit
testing of that once-monolithic system. In doing this, the help dialog is moved
up to a custom button in the UI instead of being shown in the analysis area as
was done previously.

Review URL: https://chromiumcodereview.appspot.com/10161025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133535 0039d316-1c4b-4281-b951-d872f2087c98