Double-check safe-browsing database validity on update failure.
A user reported a case where their safe-browsing database was not
updating.  Corrupt data in the add_chunks and sub_chunks area was
causing a 413 server response (even though the post was not
excessively large).  This change explicitly checks the checksums,
causing invalid databases to be reset.

This case could happen if the user's computer crashes before a
previous update was entirely committed to disk.  In most cases,
finalizing the next update would detect the corrupt data, but in this
case he error response from the server prevented the corruption from
being detected.

BUG=120219
TEST=none

Review URL: http://codereview.chromium.org/10093004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133618 0039d316-1c4b-4281-b951-d872f2087c98