Upstream WebMediaPlayerAndroid as WebKit::WebMediaPlayer implementation on android.
I've seperated the WebMediaPlayerAndroid into several changes so that each change will be easier to understand.
This is the first change in the series. It implements the basic functionalities needed for WebMediaPlayer on android.

Here are the things that are missing and will be addressed in future changes:
1. Only audio is working after this change (for video, you can only hear audio now). Video will be in the next change.
2. Fullscreen implementation is missing in this change, it will come after the video change.

BUG=
TEST=


Review URL: http://codereview.chromium.org/10073016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@133971 0039d316-1c4b-4281-b951-d872f2087c98