Pyauto-based memory test for ChromeOS in perf.py no longer freshly logs in.
When perf.MemoryTest used to log out/in (for a fresh ChromeOS login), this
caused a problem running the test via autotest because the log out cannot
be done as user chronos.  However, in discussions with nirnimesh@ and
sonnyrao@, the agreement is that there's no strong need for a fresh login
anyway in this test.  This has now been removed from the test.

BUG=chromium-os:29815
TEST=None

Review URL: https://chromiumcodereview.appspot.com/10273015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@134598 0039d316-1c4b-4281-b951-d872f2087c98