Omnibox: Add logging of what type of page was displayed when it's used
For UMA opted-in users, this adds a new field to the omnibox event record that's uploaded when the user uses the omnibox.  In particular, it records a simple enum of what type of page the user was looking at, such as a generic web page or the new tab page.

This change only adds the field to the protocol buffer upload format, not the XML format.  If I decide later I want to add it to the XML format, I'll do it in another changelist.

BUG=
TEST=using VLOG(1) of what would be uploaded


Review URL: http://codereview.chromium.org/10273019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@134824 0039d316-1c4b-4281-b951-d872f2087c98