Make the Geoposition helper struct public
This CL moves the Geoposition struct from content/common to public/content/common.h so that it can be shared by content and browser. The struct is placed in the |content| namespace and as required by the style guide, all methods except the constructor and Validate() are removed.
This paves the way for a follow-up CL in which the content will provide geolocation information to the browser.

BUG=chromium-os:22036
TEST=Chrome and tests build, run

Review URL: http://codereview.chromium.org/10316007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@134989 0039d316-1c4b-4281-b951-d872f2087c98