Change the low watermark for media buffering from 50% of buffer capacity to 90% of buffer capacity.
This is admittedly a bandaid of a fix for bug 124719 until we can think of something more clever. In short, I don't see a strong reason as to why we should cut things so close by keeping our low watermark at 50% of buffer capacity. For instance, both the browser and renderer processes have plenty of network data buffered so undeferring typically results in emptying those buffers before it trickles down to the network layer.

BUG=124719
TEST=the buffer bar should advance more regularly during playback


Review URL: http://codereview.chromium.org/10310010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135436 0039d316-1c4b-4281-b951-d872f2087c98