Remove bogus include path from ppapi/ppapi_proxy.gypi
This path is to the parent of the top-level src/ directory and was required to
support nacl. It was added to ppapi/ppapi.gyp in
http://src.chromium.org/viewvc/chrome?view=rev&revision=65614 and later moved
to ppapi/ppapi_proxy.gypi.

Use of a path outside the top-level src/ directory seems wrong. Furthermore, it
is no longer required to build either Chromium or nacl, so should be removed.

TEST=Chromium and nacl build OK


Review URL: http://codereview.chromium.org/10365020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@135462 0039d316-1c4b-4281-b951-d872f2087c98