Extract executeScript-like functionality into a single ExtensionScriptExecutorImpl class, and convert all but one* existing uses of ExtensionMsg_ExecuteCode_Params to use it.
This is so that we can extend the implementation of ExtensionScriptExecutor to notify
the UI when extensions have executed scripts, and potentially have control over whether
scripts are actually executed.

* the file not converted is chrome/browser/chromeos/accessibility/accessibility_util.cc
since it has its own script loading logic and I don't feel the need to change that.
Plus, we probably don't want any special script execution logic applying to the loading
of chromevox.

BUG=127988

Review URL: https://chromiumcodereview.appspot.com/10383104

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@136831 0039d316-1c4b-4281-b951-d872f2087c98