Browser Plugin: browser process side changes
Major classes:

BrowserPluginHost: This is the host-side of both the guest and the embedder
renderers. The embedder BrowserPluginHost keeps track of its guests, and cleans
them up when it navigates away, or whe tabs are switched or when it crashes. The
guest BrowserPluginHost knows about its embdder, and about its instance ID, as well
as some initialization parameters such as URL and initial size.

BUG=117897
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/10377170

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@138475 0039d316-1c4b-4281-b951-d872f2087c98