Pass the global matrix to CG always. Draw the minimum.
For both the cases when the canvas' device is
bitmap-backed and when it is not, pass the
global matrix to CG. When their was no backing
bitmap, set the global matrix to identity (by
concatenating the inverse).

This allows multipage documents that adjust
the global matrix by the height of the page to
draw the bitmap that acts as a proxy for the
device bits at the correct offset.

Parse the built bitmap to find the bounds of
the changed pixels, and draw that subset.

BUG=121752
BUG=129527
TEST=http://jsfiddle.net/kqxBx/3/show/ (print to PDF)
TEST=http://mrdoob.com/lab/javascript/checkbox_painter/ (print to PDF)
TEST=SkiaUtilsMacTest.BitLocker_*

Review URL: https://chromiumcodereview.appspot.com/10450024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@139304 0039d316-1c4b-4281-b951-d872f2087c98