Die, |that|, die!
This CL cleans up some uses of the "var that = this" paradigm that can be more elegently and efficiently expressed using the |bind| function. Specifically:

var that = this;
f(function(x, y, z) { that.g(x, y, z); });

Is equivalent to:

f(this.g.bind(this));

The latter is shorter, doesn't need explicit JSDoc type annotation, and is more efficient (though I doubt we'll notice any speed-up).

I haven't changed the code to use |bind| where the situation is significantly more complex than the snippet shown above. In theory it would be possible, but I doubt the code would be more readable as a result.

BUG=None
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/10444041

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@139355 0039d316-1c4b-4281-b951-d872f2087c98