Add in UMA metrics for ScreenLock requests
We are interested in knowing information about how often d-bus messages fail
when trying to lock the screen. This CL adds in a new metric,
LockScreen.RequestScreenLock, to track if sending a lock screen request
succeeded or failed. It also refactors the sending of this message to have two
distinct handling methods, so that future work with this code will be
simplier. This should give us detailed information about on what versions and
rates we are seeing this issue. It is not possible to get this information using
a few test images, like we have been using so far.

BUG=chromium-os:31346
TEST=Built Chrome with CL.
     Built ChromeOS image using CL.
     Ran image to confirm that no regressions had been introduced.
     Ran image and locked screen. Checked histogram populated.


Review URL: https://chromiumcodereview.appspot.com/10447123

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@140308 0039d316-1c4b-4281-b951-d872f2087c98