Added support for SECCOMP_RET_TRAP handlers.
When setting a sandbox policy, the user has to write a system call evaluator
function. This function is passed a system call number and returns a suitable
ErrorCode (e.g. an "errno" value).

This change list extends ErrorCode, so that in addition to static "errno" values,
the system call evaluator can also request that a callback gets called. This allows
the sandbox to handle system calls in user space.

BUG=130662
TEST=make && ./demo32 && ./demo64

Review URL: https://chromiumcodereview.appspot.com/10533076

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142353 0039d316-1c4b-4281-b951-d872f2087c98