Re-enable regression test for packing and installing extensions on Windows.
I changed it to wait for NOTIFICATION_CRX_INSTALLER_DONE as that notification is
definitely sent after confirmation_requested is set to true. Before, it waited
for NOTIFICATION_EXTENSION_INSTALLED which occurs at about the same time,
causing the test to be flaky for Windows.

BUG=130951
TEST=Yes


Review URL: https://chromiumcodereview.appspot.com/10540068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142457 0039d316-1c4b-4281-b951-d872f2087c98