Revert 142611 - Return FileEntry object from fileBrowserHandler.selectFile function.
Broke FileSystemExtensionApi tests.

Currently, the function returns file's filesystem URL, and the caller has to resolve it to get the entry.

It would be more convenient for the function caller it the file entry is returned in the first place.
Also this is more in line with onExecute event which dispatches an array of entries.

BUG=131716
TEST=*FileBrowserHandler*

Review URL: https://chromiumcodereview.appspot.com/10535051

TBR=tbarzic@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10532185

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142614 0039d316-1c4b-4281-b951-d872f2087c98