Fixed bug causing invalid PPAPI test to PASS.
If a test was run with a filter that did not name a valid test, no test would be run and success would be reported. This is bad because it allows a test to be added to ppapi_uitest.cc which does not run and no error is raised.

This also fixes all existing invalid tests.

BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10627012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@143655 0039d316-1c4b-4281-b951-d872f2087c98