Aura: Use SyncPoints instead of waiting for OnCompositingEnded
When an new frame comes in from a renderer, we do a pointer flip from an old to a new renderer texture. Previously, We waited until the next UI frame cleared (via OnCompositingEnded) before ACKing the renderer. This ensured that there were no pending UI GL calls in the pipeline referring to the old renderer texture.

With this CL, we ACK the renderer once we have kicked the frame (a form of flow control ensuring that the renderer does not race ahead of the UI) and use a SyncPoint in the GPU process to make sure all the pipelined UI draw calls have cleared before processing more renderer draw calls. 

TEST=none
BUG=112299


Review URL: https://chromiumcodereview.appspot.com/10576015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@144053 0039d316-1c4b-4281-b951-d872f2087c98