Use ICU script code "Jpan" instead of "Hrkt" in Japanese pref names.
The script code "Jpan" ("Japanese") is more suitable for naming font prefs for rendering Japanese than "Hrkt" ("Hiragana or Katakana").

With this change, the script code used in Chrome pref names for Japanese will become different than the one used in WebKit settings. We would have used "Jpan" in both Chrome and WebKit to begin with, but it is a newer script code that the version of ICU included in WebKit for MacOS doesn't have. But since Chrome now exposes script codes in the Font Settings Extension API, it's better to use the more appropriate, modern one. This patch just changes the script used internally by Chrome and translates it to the one used by WebKit when talking to WebCore.

BUG=2685
TEST=browser_tests --gtest_filter=ExtensionApiTest.FontSetting*
TBR=pkasting,mpcomplete

Review URL: https://chromiumcodereview.appspot.com/10532105

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@144418 0039d316-1c4b-4281-b951-d872f2087c98